+++ subsys/bluetooth/controller/include/ll.h
@@ -66,7 +66,7 @@ u8_t ll_priv_mode_set(bt_addr_le_t *id_addr, u8_t mode);
 
 u8_t ll_create_connection(u16_t scan_interval, u16_t scan_window,
 			  u8_t filter_policy, u8_t peer_addr_type,
-			  u8_t *p_peer_addr, u8_t own_addr_type,
+			  u8_t *peer_addr, u8_t own_addr_type,
 			  u16_t interval, u16_t latency, u16_t timeout);
 u8_t ll_connect_disable(void **rx);
 u8_t ll_conn_update(u16_t handle, u8_t cmd, u8_t status, u16_t interval_min,
