+++ subsys/bluetooth/controller/ll_sw/nordic/lll/lll_adv.h
@@ -34,6 +34,10 @@ struct lll_adv {
 	u8_t is_mesh:1;
 #endif /* CONFIG_BT_HCI_MESH_EXT */
 
+#if defined(CONFIG_BT_CTLR_PRIVACY)
+	u8_t  rl_idx;
+#endif /* CONFIG_BT_CTLR_PRIVACY */
+
 	struct lll_adv_pdu adv_data;
 	struct lll_adv_pdu scan_rsp;
 };
